renaming some folders + starting to test compile code

+++ scan/src/arp/if_setup.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 
-int maskToPrefix(uint32_t subMaskBin);
+uint8_t maskToPrefix(uint32_t subMaskBin);
 int getCommandOutput(const char *command, char *buf);
 const char *command ="nmcli -t -f DEVICE,TYPE,STATE dev status | awk -F: '$2==\"wifi\" && $3==\"connected\"{print $1; exit}'";
 int ifIndex(char *ssid);
